[CL 2744667 by Andrew Grant in Main branch]

+++ Engine/Source/Developer/FriendsAndChat/Public/FriendsChatStyle.h
@@ -7,7 +7,7 @@
 // Forward declarations
 namespace EChatMessageType
 {
-	enum Type : uint8;
+	enum Type : uint16;
 }
 
 /**
